+++ examples/H5Part/read_core_vfd.c
@@ -45,7 +45,7 @@ main (
 
 	// with core cfd we read the hole file on all cores!
         h5_int64_t num_particles = H5PartGetNumParticles (file);
-        printf ("[proc %d]: particles in view: %lld\n", comm_rank, num_particles);
+        printf ("[proc %d]: particles in view: %lld\n", comm_rank, (long long)num_particles);
 
 	// read and print data
         h5_int32_t* data = calloc (num_particles, sizeof (*data));
